requirements.txt
beautifulsoup4==4.6.0
Django==1.10.6
django-blog-zinnia==0.19.1.dev0
django-contrib-comments==1.8.0
django-haystack==2.7.0
django-js-asset==1.0.0
django-mptt==0.9.0
django-tagging==0.4.6
django-xmlrpc==0.1.8
jieba==0.39
Markdown==2.6.8
mots-vides==2015.5.11
Pillow==5.0.0
Pygments==2.2.0
pyparsing==2.2.0
pytz==2017.2
regex==2018.2.8
Whoosh==2.7.4
